Create Flattening a linked list
Flattening a given linked list where all the nodes have a next pointer and a bottom pointer. First we used recursion to flatten all the nodes based on their bottom pointer except the first node and its' bottom pointers. Once recursion returned us the flattened version of all the nodes except the first one, we merged both the flattened list as well as the list with first node and its' bottom pointers. Then return the MergeList.